move save_config in automl.common.util to chronos.feature.utils
